+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (New) ValidationResult: allow accumulating errors.
 - (New) When we report an unhandled exception, filter some stackframes that are not relevant to the user code. Full stacktrace still logged at trace level.
 - (Breaking) Due to changes in how objectives are handled, ReferenceResult methods have been renamed for clarity.
+- (Breaking) Removed Improver::_improve, please implement Improver::improve directly instead. To migrate, just rename the method and make it public.
 - (Fix) Math.random, Collections.shuffle now blocked using AspectJ instead of reflection. --add-opens no longer necessary.
 - (Fix) Add @JsonIgnore to instance in solution class already exists in getter, but serializer may be configured to use fields instead of methods.
 - (Fix) Gracefully handle experiment declarations with 0 algorithms.